refactor(components): add `ui/` subdirectory
