+++ src/components/ui/Display/Display.module.scss
