+++ src/components/ui/Display/interfaces.ts
